Filter func useable with pd structures;neutral labels
Trying out using negative label values for silence preceding a spoken
word. So the label for the silence before the 1st word (labeled 1) is
labeled -1, for 2nd word it is -2, and so on. To get all silence, fitler
to below zero. This helps make sure data from silent regions is
contigous. But this make cause more confusion than it solves