Random label target transform; easier sensor config
Use RandomIntLike to sample a uniform random integer the same size as input. Use to do
even splits across classes in the same shape as the original target
array.

Determine which sensors to use based on the electrodes mat array. By
default, the sensors required during training are passed as required
sensors for CV and testing sets, so missing/bads should be imputed as
zeros in these cases.